Autocomplete: extract the fast-path client (#5284)
