+++ vscode/src/completions/default-client.ts
